+++ tests/test_readme.py
@@ -13,10 +13,21 @@ class TestReadme(unittest.TestCase):
 
         # Create a new HDF5 file with FAIR metadata
         with h5tbx.File("example.h5", "w") as h5:
-            h5.create_dataset("temperature", data=np.array([20, 21, 19, 22]))
-            h5.attrs["units", M4I.hasUnit] = "degree_Celsius"
-            h5.rdf["units"].object = "http://qudt.org/vocab/unit/DEG_C"
-            h5.attrs["description", "https://schema.org/description"] = "Room temperature measurements"
+            ds = h5.create_dataset("temperature", data=np.array([20, 21, 19, 22]))
+            ds.attrs["units"] = h5tbx.Attribute(
+                value="degree_Celsius",
+                rdf_predicate=M4I.hasUnit,
+                rdf_object="http://qudt.org/vocab/unit/DEG_C",
+            )
+            ds.attrs["description", "https://schema.org/description"] = "Room temperature measurements"
 
-            ttl = h5.serialize("ttl")
-        print(ttl)
+            ds_mean = h5.create_dataset("mean_temperature", data=np.mean(h5["temperature"][()]))
+            ds_mean.attrs["units", M4I.hasUnit] = "degree_Celsius"
+            ds_mean.rdf["units"].object = "http://qudt.org/vocab/unit/DEG_C"
+            ds_mean.attrs["description", "https://schema.org/description"] = "Mean room temperature measurements"
+
+            ds_mean.rdf.type = M4I.NumericalVariable
+            ds_mean.rdf.data_predicate = M4I.hasNumericalValue
+
+            ttl = h5.serialize("ttl", structural=True, contextual=False, file_uri="https://example.org#example.h5/")
+            print(ttl)